Refactor/separate exceptions into a separate file (#169)
* refactor: separate exception classes into a separate file

* refactor: modify the import source

* docs: remove unused noqa
