+++ src/majsoulrpa/_rpa.py
@@ -15,11 +15,11 @@ from ._impl.zmq_client import ZMQClient
 from .common import timeout_to_deadline
 from .presentation import AuthPresentation, HomePresentation, LoginPresentation
 from .presentation._presentation_creator import PresentationCreator
-from .presentation.presentation_base import (
-    PresentationBase,
+from .presentation.exceptions import (
     PresentationNotDetectedError,
     PresentationTimeoutError,
 )
+from .presentation.presentation_base import PresentationBase
 
 if TYPE_CHECKING:
     from ._impl.message_queue_client import MessageQueueClientBase
