+++ src/majsoulrpa/presentation/auth.py
@@ -8,13 +8,15 @@ from majsoulrpa._impl.message_queue_client import MessageQueueClientBase
 from majsoulrpa._impl.template import Template
 from majsoulrpa.common import TimeoutType, timeout_to_deadline
 
-from .presentation_base import (
+from .exceptions import (
     InvalidOperationError,
+    PresentationNotDetectedError,
+    PresentationTimeoutError,
+)
+from .presentation_base import (
     Presentation,
     PresentationBase,
     PresentationCreatorBase,
-    PresentationNotDetectedError,
-    PresentationTimeoutError,
 )
 
 _MAX_EMAIL_ADDRESS_LENGTH: Final[int] = 50  # JP ver
